requirements.txt
Flask==2.3.3
boto3==1.34.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.32.4
 beautifulsoup4==4.13.4
 vosk==0.3.45
